refactor: regroup pages into a folder
